downloads interaction service
